added sync and refactored sync methods to new class. (fixes #49)
