+++ app/src/main/java/org/ole/planet/takeout/LoginActivity.java
@@ -42,16 +42,14 @@ import java.util.Arrays;
 import java.util.List;
 
 
-public class LoginActivity extends AppCompatActivity {
+public class LoginActivity extends SyncActivity {
     private EditText inputName, inputPassword;
     private TextInputLayout inputLayoutName, inputLayoutPassword;
     private Button btnSignIn;
     private ImageButton imgBtnSetting;
-    private TextView syncOption, syncDate, intervalLabel;
-    private Spinner spinner;
-    private Switch syncSwitch;
+    private TextView syncOption;
     Context context;
-    private View positiveAction, positiveSync, negativeSync;
+    private View positiveAction;
     boolean connectionResult;
     dbSetup dbsetup =  new dbSetup();
 
@@ -198,98 +196,6 @@ public class LoginActivity extends AppCompatActivity {
         dialog.show();
     }
 
-
-    // Server feedback dialog
-    public void  feedbackDialog(){
-
-        MaterialDialog dialog = new MaterialDialog.Builder(this)
-                .title(R.string.title_sync_settings)
-                .customView(R.layout.dialog_sync_feedback, true)
-                .positiveText(R.string.btn_sync)
-                .negativeText(R.string.btn_sync_cancel)
-                .neutralText(R.string.btn_sync_save).onPositive(new MaterialDialog.SingleButtonCallback() {
-                    @Override
-                    public void onClick(MaterialDialog dialog, DialogAction which) {
-                        Log.e("MD: ", "Clicked Positive (Sync now)");
-                        Toast.makeText(LoginActivity.this, "Syncing now...", Toast.LENGTH_SHORT).show();
-                    }
-                })
-                .onNegative(new MaterialDialog.SingleButtonCallback() {
-                    @Override
-                    public void onClick(MaterialDialog dialog, DialogAction which) {
-                        Log.e("MD: ", "Clicked Negative (Cancel)");
-                    }
-                })
-                .onNeutral(new MaterialDialog.SingleButtonCallback() {
-                    @Override
-                    public void onClick(@NonNull MaterialDialog dialog, @NonNull DialogAction which) {
-                        Log.e("MD: ", "Clicked Netural (Save)");
-                        Toast.makeText(LoginActivity.this, "Saving sync settings...", Toast.LENGTH_SHORT).show();
-                    }
-                })
-                .build();
-
-        // Check Autosync switch (Toggler)
-        syncSwitch = (Switch) dialog.findViewById(R.id.syncSwitch);
-        intervalLabel = (TextView) dialog.findViewById(R.id.intervalLabel);
-
-
-        syncSwitch.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                if (isChecked) {
-                    Log.e("MD: ", "Autosync is On");
-                    intervalLabel.setVisibility(View.VISIBLE);
-                    spinner.setVisibility(View.VISIBLE);
-                } else {
-                    Log.e("MD: ", "Autosync is Off");
-                    spinner.setVisibility(View.GONE);
-                    intervalLabel.setVisibility(View.GONE);
-                }
-            }
-        });
-
-        int convertedDate = convertDate();
-
-        // Check if the user never synced
-        if (convertedDate == 0){
-            syncDate = (TextView) dialog.findViewById(R.id.lastDateSynced);
-            syncDate.setText("Last Sync Date: Never");
-        }
-        else {
-            syncDate = (TextView) dialog.findViewById(R.id.lastDateSynced);
-            syncDate.setText("Last Sync Date: " + convertedDate);
-        }
-
-        // Init spinner dropdown items
-        spinner = (Spinner) dialog.findViewById(R.id.intervalDropper);
-        syncDropdownAdd();
-
-        dialog.show();
-    }
-
-    // Converts OS date to human date
-    private int convertDate(){
-        // Context goes here
-
-        return 0; // <=== modify this when implementing this method
-    }
-    
-    // Create items in the spinner
-    public void syncDropdownAdd(){
-        List<String> list = new ArrayList<>();
-
-        list.add("15 Minutes");
-        list.add("30 Minutes");
-        list.add("1 Hour");
-        list.add("3 Hours");
-
-        ArrayAdapter<String> adapter = new ArrayAdapter<String>(getApplicationContext(),
-                android.R.layout.simple_spinner_item, list);
-        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-        spinner.setAdapter(adapter);
-
-    }
 
     public boolean isServerReachable(String url) {
         final Fuel ful = new Fuel();
